requirements.txt
# Nell Podcast Enhancement System - Requirements
# Author: Sarath

# Core dependencies
anthropic>=0.18.0
fal-client>=0.3.0
python-dotenv>=1.0.0

# Audio processing
pydub>=0.25.1
openai-whisper>=20231117  # For audio/video transcription

# Multi-format input support
pypdf>=4.0.0              # PDF extraction
python-docx>=1.0.0        # Word document extraction
newspaper3k>=0.2.8        # URL/article extraction
moviepy>=1.0.3            # Video processing

# HTTP requests
requests>=2.31.0
aiohttp>=3.9.0            # Async HTTP for parallel execution

# Configuration
pyyaml>=6.0

# FFmpeg (system dependency - install separately)
# brew install ffmpeg (macOS)
# apt install ffmpeg (Ubuntu)
